+++ devices.js
@@ -1,5 +1,9 @@
-const xymd = require('./devices/xy-md-02');
+const xymd = require('./devices/XY-MD02');
+const sdm120 = require('./devices/SDM120');
 
-const devices = [xymd];
+const devices = [
+  xymd,
+  sdm120,
+];
 
 module.exports = devices;